+++ skeleton/Ship.hpp
@@ -10,7 +10,7 @@ public:
 	virtual void step(double dt) override;
 	virtual void handle_keyboard_button_down(unsigned char c) override;
 	virtual void handle_keyboard_button_up(unsigned char c) override;
-	virtual void handle_mouse_pos(int x, int y) override;
+	virtual void handle_mouse_pos(float x, float y) override;
 protected:
 	struct angular_velocity {
 		float angle;
